ErrorBoundary to new file
